cleanup modal, missing icon
